Fixed override and comparicine issues

+++ Minigin/GamePadSystem.h
@@ -6,6 +6,6 @@ namespace dae {
 	public:
 		virtual ~GamePadSystem() = default;
 		virtual void Initialize() = 0;
-		virtual int GetPlayerIdFromIndex(int index) const = 0;
+		virtual int GetPlayerIdFromIndex(size_t index) const = 0;
 	};
 }
